refactor: move event matching code to types/testutil
