+++ types/testutil/events.go
@@ -1,4 +1,4 @@
-package utils
+package testutil
 
 import (
 	"encoding/json"
