+++ x/ecocredit/server/core/msg_update_class_admin_test.go
@@ -11,8 +11,8 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	api "github.com/regen-network/regen-ledger/api/regen/ecocredit/v1"
+	"github.com/regen-network/regen-ledger/types/testutil"
 	"github.com/regen-network/regen-ledger/x/ecocredit/core"
-	"github.com/regen-network/regen-ledger/x/ecocredit/server/utils"
 )
 
 type updateClassAdmin struct {
@@ -101,9 +101,9 @@ func (s *updateClassAdmin) ExpectEventWithProperties(a gocuke.DocString) {
 	err := json.Unmarshal([]byte(a.Content), &event)
 	require.NoError(s.t, err)
 
-	sdkEvent, found := utils.GetEvent(&event, s.sdkCtx.EventManager().Events())
+	sdkEvent, found := testutil.GetEvent(&event, s.sdkCtx.EventManager().Events())
 	require.True(s.t, found)
 
-	err = utils.MatchEvent(&event, sdkEvent)
+	err = testutil.MatchEvent(&event, sdkEvent)
 	require.NoError(s.t, err)
 }
